+++ Assets/Scripts/WorkManager.cs
@@ -98,7 +98,7 @@ public class WorkManager : MonoBehaviour
 	{
 		if(gameObject.GetComponent<GameManager>().money >= buyoutCost) {
 			gameObject.GetComponent<GameManager>().money -= buyoutCost;
-			AddClicks(currentJob.ClicksToPromotion - currentClicks);
+			AddClicks((int)currentJob.ClicksToPromotion - currentClicks);
 		} else
 			Debug.Log("Not enough money to buy promotion!");
 	}
